+++ packages/coffee-bot-board/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @google-labs/coffee-bot-board
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [e8d0737]
+  - @google-labs/breadboard@0.14.0
+  - @google-labs/core-kit@0.5.2
+  - @google-labs/json-kit@0.1.3
+  - @google-labs/node-nursery@0.0.9
+  - @google-labs/palm-kit@0.0.8
+  - @google-labs/template-kit@0.2.3
+
 ## 0.0.8
 
 ### Patch Changes