+++ packages/node-nursery-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @google-labs/node-nursery-web
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [e8d0737]
+  - @google-labs/breadboard@0.14.0
+
 ## 1.0.5
 
 ### Patch Changes